--- a/evm/src/main/java/org/hyperledger/besu/evm/precompile/BigIntegerModularExponentiationPrecompiledContract.java
+++ b/evm/src/main/java/org/hyperledger/besu/evm/precompile/BigIntegerModularExponentiationPrecompiledContract.java
@@ -123,8 +123,7 @@ public class BigIntegerModularExponentiationPrecompiledContract
       final int baseLength = clampedToInt(length_of_BASE);
       final int modulusOffset = clampedToInt(BASE_OFFSET + length_of_BASE + length_of_EXPONENT);
       final int modulusLength = clampedToInt(length_of_MODULUS);
-      if ((extractLastByte(input, baseOffset, baseLength) & 1) != 1
-          && (extractLastByte(input, modulusOffset, modulusLength) & 1) != 1) {
+      if ((extractLastByte(input, baseOffset, baseLength) & 1) != 1) {
         return computeNative(input, modulusLength);
       }
     }