--- a/evm/src/main/java/org/hyperledger/besu/evm/precompile/BigIntegerModularExponentiationPrecompiledContract.java
+++ b/evm/src/main/java/org/hyperledger/besu/evm/precompile/BigIntegerModularExponentiationPrecompiledContract.java
@@ -121,7 +121,6 @@ public class BigIntegerModularExponentiationPrecompiledContract
     if (LibArithmetic.ENABLED) {
       final int baseOffset = clampedToInt(BASE_OFFSET);
       final int baseLength = clampedToInt(length_of_BASE);
-      final int modulusOffset = clampedToInt(BASE_OFFSET + length_of_BASE + length_of_EXPONENT);
       final int modulusLength = clampedToInt(length_of_MODULUS);
       if ((extractLastByte(input, baseOffset, baseLength) & 1) != 1) {
         return computeNative(input, modulusLength);